bots_modeling/requirements.txt
openai
python-dotenv
pymongo
matplotlib
scikit-learn
seaborn
requests
